fix: address CodeRabbit PR comments
- Remove unused behavioralAgentErrorSchema from agent-behavioral.ts
- Fix invalid JSON examples (0-100 -> numeric values)
- Add required fields to response schema in ai.routes.ts
- Fix TOCTOU race in rate limiting (combined checkAndIncrement)
- Reorder operations (increment before save)
- Extract duplicate Gemini code to gemini-client.ts
- Fix API key in header instead of URL
- Fix retry loop condition and timer cleanup
